+++ src/Castle.Core/DynamicProxy/Generators/MethodSignatureComparer.cs
@@ -22,6 +22,8 @@ namespace Castle.DynamicProxy.Generators
 	{
 		public static readonly MethodSignatureComparer Instance = new MethodSignatureComparer();
 
+		private static readonly Type preserveBaseOverridesAttribute = Type.GetType("System.Runtime.CompilerServices.PreserveBaseOverridesAttribute", throwOnError: false);
+
 		public bool EqualGenericParameters(MethodInfo x, MethodInfo y)
 		{
 			if (x.IsGenericMethod != y.IsGenericMethod)
@@ -77,7 +79,12 @@ namespace Castle.DynamicProxy.Generators
 			return true;
 		}
 
-		public bool EqualSignatureTypes(Type x, Type y)
+		public bool EqualReturnTypes(MethodInfo x, MethodInfo y)
+		{
+			return EqualSignatureTypes(x.ReturnType, y.ReturnType, x, y);
+		}
+
+		private bool EqualSignatureTypes(Type x, Type y, MethodInfo xm = null, MethodInfo ym = null)
 		{
 			if (x.IsGenericParameter != y.IsGenericParameter)
 			{
@@ -122,6 +129,15 @@ namespace Castle.DynamicProxy.Generators
 			{
 				if (!x.Equals(y))
 				{
+					// This enables covariant method returns for .NET 5 and newer.
+					// No need to check for runtime support, since such methods are marked with a custom attribute;
+					// see https://github.com/dotnet/runtime/blob/main/docs/design/features/covariant-return-methods.md.
+					if (preserveBaseOverridesAttribute != null)
+					{
+						return (xm != null && xm.IsDefined(preserveBaseOverridesAttribute, inherit: false) && y.IsAssignableFrom(x))
+						    || (ym != null && ym.IsDefined(preserveBaseOverridesAttribute, inherit: false) && x.IsAssignableFrom(y));
+					}
+
 					return false;
 				}
 			}
@@ -142,7 +158,7 @@ namespace Castle.DynamicProxy.Generators
 
 			return EqualNames(x, y) &&
 				   EqualGenericParameters(x, y) &&
-				   EqualSignatureTypes(x.ReturnType, y.ReturnType) &&
+				   EqualReturnTypes(x, y) &&
 				   EqualParameters(x, y);
 		}